+++ igzip/adler32_base.c
@@ -60,5 +60,5 @@ adler32_base(uint32_t adler32, uint8_t *start, uint64_t length)
         A = A % ADLER_MOD;
         B = B % ADLER_MOD;
 
-        return B << 16 | A;
+        return (uint32_t) ((B << 16) | A);
 }
